refactor database and update API

+++ internal/app/service/contract_service.go
@@ -6,9 +6,10 @@ import (
 	"awesomeProject/internal/app/domain/dao"
 	"awesomeProject/internal/app/pkg"
 	"awesomeProject/internal/app/repository"
-	"fmt"
 	"github.com/gin-gonic/gin"
+	"github.com/google/uuid"
 	"net/http"
+	"os"
 	"strconv"
 	"time"
 )
@@ -19,7 +20,6 @@ type ContractService interface {
 	Create(c *gin.Context)
 	Update(c *gin.Context)
 	Delete(c *gin.Context)
-	GetAllByRenterOrLessorID(c *gin.Context)
 }
 
 type ContractServiceImpl struct {
@@ -27,6 +27,8 @@ type ContractServiceImpl struct {
 	hashContractRepo   repository.HashContractRepository
 	servicesDemandRepo repository.ServicesDemandRepository
 	invoiceRepo        repository.InvoiceRepository
+	roomRepo           repository.RoomRepository
+	userRepo           repository.UserRepository
 }
 
 type ContractParams struct {
@@ -40,22 +42,57 @@ type ContractParams struct {
 	Status             int       `json:"status"`
 	IsEnable           bool      `json:"is_enable"`
 	FileBase64         string    `json:"file_base64"`
+	FileName           string    `json:"file_name"`
 	ChargeableServices []int     `json:"chargeable_services"`
+	IsRenterSigned     bool      `json:"is_renter_signed"`
+	IsLessorSigned     bool      `json:"is_lessor_signed"`
+	MonthlyPrice       float64   `json:"monthly_price"`
 }
 
 func (repo ContractServiceImpl) GetAll(c *gin.Context) {
-	data, err := repo.contractRepo.GetAll()
-	fmt.Println("test")
+	renterID, _ := strconv.Atoi(c.Query("renter_id"))
+	lessorID, _ := strconv.Atoi(c.Query("lessor_id"))
 
-	contract := blockchain.InvokeChaincode()
-	blockchain.GetAllAssets(contract)
+	filter := &repository.ContractFilter{
+		RenterID: renterID,
+		LessorID: lessorID,
+	}
+
+	data, err := repo.contractRepo.GetAll(filter)
+
+	//contract := blockchain.InvokeChaincode()
+	//blockchain.GetAllAssets(contract)
 
 	if err != nil {
 		c.JSON(http.StatusBadRequest, pkg.BuildResponse(constant.BadRequest, pkg.Null(), err))
 		return
 	}
 
-	c.JSON(http.StatusOK, pkg.BuildResponse(constant.Success, pkg.Null(), data))
+	var contracts []dao.ContractResponse
+	for _, contract := range data {
+		renter, _ := repo.userRepo.GetByID(int(contract.RenterID))
+		lessor, _ := repo.userRepo.GetByID(int(contract.LessorID))
+		room, _ := repo.roomRepo.GetByID(int(contract.RoomID))
+		contracts = append(contracts, dao.ContractResponse{
+			ID:            contract.ID,
+			Renter:        *renter,
+			Lessor:        *lessor,
+			Room:          *room,
+			MonthlyPrice:  contract.MonthlyPrice,
+			CanceledBy:    nil,
+			DateRent:      contract.DateRent,
+			DatePay:       contract.DatePay,
+			PayMode:       contract.PayMode,
+			Payment:       contract.Payment,
+			Status:        contract.Status,
+			IsEnable:      contract.IsEnable,
+			FilePath:      contract.FilePath,
+			Invoices:      contract.Invoices,
+			ServiceDemand: contract.ServiceDemands,
+		})
+	}
+
+	c.JSON(http.StatusOK, pkg.BuildResponse(constant.Success, pkg.Null(), contracts))
 }
 
 func (repo ContractServiceImpl) GetByID(c *gin.Context) {
@@ -66,34 +103,57 @@ func (repo ContractServiceImpl) GetByID(c *gin.Context) {
 		return
 	}
 
-	contractBC := blockchain.InvokeChaincode()
-	_, err = blockchain.ReadAsset(contractBC, int64(id))
-	if err != nil {
-		c.JSON(http.StatusBadRequest, pkg.BuildResponse(constant.BadRequest, pkg.Null(), err))
-		return
+	if data.IsLessorSigned && data.IsRenterSigned {
+		contractBC := blockchain.InvokeChaincode()
+		_, err = blockchain.ReadAsset(contractBC, int64(id))
+		if err != nil {
+			c.JSON(http.StatusBadRequest, pkg.BuildResponse(constant.BadRequest, pkg.Null(), err))
+			return
+		}
+
+		base64, err := pkg.ConvertFileToBase64(data.FilePath)
+		if err != nil {
+			return
+		}
+
+		fileHashed, _ := pkg.HashFileBase64ToSHA256(base64)
+
+		hashContractData, err := repo.hashContractRepo.GetByContractID(int(data.ID))
+
+		if hashContractData == nil {
+			c.JSON(http.StatusBadRequest, pkg.BuildResponse(constant.BadRequest, pkg.Null(), "Hash contract not found"))
+			return
+		}
+		hashContract, err := blockchain.ReadHashContract(contractBC, int64(hashContractData.ID))
+
+		if hashContract != nil && fileHashed != hashContract.Hash {
+			c.JSON(http.StatusBadRequest, pkg.BuildResponse(constant.BadRequest, pkg.Null(), "File is modified"))
+			return
+		}
 	}
 
-	base64, err := pkg.ConvertFileToBase64(data.FilePath)
-	if err != nil {
-		return
+	renter, _ := repo.userRepo.GetByID(int(data.RenterID))
+	lessor, _ := repo.userRepo.GetByID(int(data.LessorID))
+	room, _ := repo.roomRepo.GetByID(int(data.RoomID))
+	contract := dao.ContractResponse{
+		ID:            data.ID,
+		Renter:        *renter,
+		Lessor:        *lessor,
+		Room:          *room,
+		MonthlyPrice:  data.MonthlyPrice,
+		CanceledBy:    nil,
+		DateRent:      data.DateRent,
+		DatePay:       data.DatePay,
+		PayMode:       data.PayMode,
+		Payment:       data.Payment,
+		Status:        data.Status,
+		IsEnable:      data.IsEnable,
+		FilePath:      data.FilePath,
+		Invoices:      data.Invoices,
+		ServiceDemand: data.ServiceDemands,
 	}
 
-	fileHashed, _ := pkg.HashFileBase64ToSHA256(base64)
-
-	hashContractData, err := repo.hashContractRepo.GetByContractID(int(data.ID))
-
-	if hashContractData == nil {
-		c.JSON(http.StatusBadRequest, pkg.BuildResponse(constant.BadRequest, pkg.Null(), "Hash contract not found"))
-		return
-	}
-	hashContract, err := blockchain.ReadHashContract(contractBC, int64(hashContractData.ID))
-
-	if hashContract != nil && fileHashed != hashContract.Hash {
-		c.JSON(http.StatusBadRequest, pkg.BuildResponse(constant.BadRequest, pkg.Null(), "File is modified"))
-		return
-	}
-
-	c.JSON(http.StatusOK, pkg.BuildResponse(constant.Success, pkg.Null(), data))
+	c.JSON(http.StatusOK, pkg.BuildResponse(constant.Success, pkg.Null(), contract))
 }
 
 func (repo ContractServiceImpl) Create(c *gin.Context) {
@@ -105,19 +165,26 @@ func (repo ContractServiceImpl) Create(c *gin.Context) {
 		return
 	}
 
-	filePath, err := pkg.SaveFile(contractDAO.FileBase64)
+	bucket, _ := os.LookupEnv("AWS_BUCKET")
+	url, err := pkg.UploadS3(bucket, "rooms/"+uuid.New().String()+contractDAO.FileName+"/", []byte(contractDAO.FileBase64))
+	if err != nil {
+		return
+	}
 
 	contract := &dao.Contract{
-		RenterID: uint(contractDAO.RenterID),
-		LessorID: uint(contractDAO.LessorID),
-		RoomID:   uint(contractDAO.RoomID),
-		DateRent: contractDAO.DateRent,
-		DatePay:  contractDAO.DatePay,
-		PayMode:  contractDAO.PayMode,
-		Payment:  contractDAO.Payment,
-		Status:   contractDAO.Status,
-		IsEnable: contractDAO.IsEnable,
-		FilePath: filePath,
+		RenterID:       uint(contractDAO.RenterID),
+		LessorID:       uint(contractDAO.LessorID),
+		RoomID:         uint(contractDAO.RoomID),
+		DateRent:       contractDAO.DateRent,
+		DatePay:        contractDAO.DatePay,
+		PayMode:        contractDAO.PayMode,
+		Payment:        contractDAO.Payment,
+		Status:         contractDAO.Status,
+		IsEnable:       contractDAO.IsEnable,
+		FilePath:       url,
+		IsRenterSigned: contractDAO.IsRenterSigned,
+		IsLessorSigned: contractDAO.IsLessorSigned,
+		MonthlyPrice:   contractDAO.MonthlyPrice,
 	}
 	data, err := repo.contractRepo.Create(contract)
 
@@ -126,27 +193,29 @@ func (repo ContractServiceImpl) Create(c *gin.Context) {
 		return
 	}
 
-	hashFile, _ := pkg.HashFileBase64ToSHA256(contractDAO.FileBase64)
-	hashContract := &dao.HashContract{
-		ContractID: data.ID,
-		Hash:       hashFile,
-	}
-	dataHash, err := repo.hashContractRepo.Create(hashContract)
+	if data.IsLessorSigned && data.IsRenterSigned {
+		hashFile, _ := pkg.HashFileBase64ToSHA256(contractDAO.FileBase64)
+		hashContract := &dao.HashContract{
+			ContractID: data.ID,
+			Hash:       hashFile,
+		}
+		dataHash, err := repo.hashContractRepo.Create(hashContract)
 
-	contractBC := blockchain.InvokeChaincode()
-	err = blockchain.CreateAssets(contractBC, data)
+		contractBC := blockchain.InvokeChaincode()
+		err = blockchain.CreateAssets(contractBC, data)
 
-	if err != nil {
-		_ = repo.contractRepo.Delete(int(data.ID))
-		c.JSON(http.StatusBadRequest, pkg.BuildResponse(constant.BadRequest, pkg.Null(), err))
-		return
-	}
+		if err != nil {
+			_ = repo.contractRepo.Delete(int(data.ID))
+			c.JSON(http.StatusBadRequest, pkg.BuildResponse(constant.BadRequest, pkg.Null(), err))
+			return
+		}
 
-	err = blockchain.CreateHashContract(contractBC, dataHash)
+		err = blockchain.CreateHashContract(contractBC, dataHash)
 
-	if err != nil {
-		c.JSON(http.StatusBadRequest, pkg.BuildResponse(constant.BadRequest, pkg.Null(), err))
-		return
+		if err != nil {
+			c.JSON(http.StatusBadRequest, pkg.BuildResponse(constant.BadRequest, pkg.Null(), err))
+			return
+		}
 	}
 
 	c.JSON(http.StatusOK, pkg.BuildResponse(constant.Success, pkg.Null(), data))
@@ -183,28 +252,19 @@ func (repo ContractServiceImpl) Delete(c *gin.Context) {
 	c.JSON(http.StatusOK, pkg.BuildResponse(constant.Success, pkg.Null(), pkg.Null()))
 }
 
-func (repo ContractServiceImpl) GetAllByRenterOrLessorID(c *gin.Context) {
-	renterID, _ := strconv.Atoi(c.Query("renter_id"))
-	lessorID, _ := strconv.Atoi(c.Query("lessor_id"))
-	data, err := repo.contractRepo.GetAllByRenterOrLessorID(renterID, lessorID)
-
-	if err != nil {
-		c.JSON(http.StatusBadRequest, pkg.BuildResponse(constant.BadRequest, pkg.Null(), err))
-		return
-	}
-
-	c.JSON(http.StatusOK, pkg.BuildResponse(constant.Success, pkg.Null(), data))
-}
-
 func ContractServiceInit(
 	repo repository.ContractRepository,
 	hashContractRepo repository.HashContractRepository,
 	servicesDemandRepo repository.ServicesDemandRepository,
-	invoiceRepo repository.InvoiceRepository) *ContractServiceImpl {
+	invoiceRepo repository.InvoiceRepository,
+	roomRepo repository.RoomRepository,
+	userRepo repository.UserRepository) *ContractServiceImpl {
 	return &ContractServiceImpl{
 		contractRepo:       repo,
 		hashContractRepo:   hashContractRepo,
 		servicesDemandRepo: servicesDemandRepo,
 		invoiceRepo:        invoiceRepo,
+		roomRepo:           roomRepo,
+		userRepo:           userRepo,
 	}
 }
